Added a CardFilter class

+++ src/main/java/io/github/cmansfield/card/Card.java
@@ -37,7 +37,7 @@ public class Card implements Serializable {
     return getFromPojo(CardConstants.MANA_COST.key());
   }
 
-  public Integer getCmc() {
+  public Double getCmc() {
     return getFromPojo(CardConstants.CMC.key());
   }
